@@ -1,6 +1,6 @@
 {
   "name": "react-bootstrap",
-  "version": "0.23.5",
+  "version": "0.24.3",
   "description": "Bootstrap 3 components build with React",
   "repository": {
     "type": "git",
@@ -43,71 +43,72 @@
     "react": ">=0.13"
   },
   "devDependencies": {
-    "babel": "^5.1.10",
-    "babel-core": "^5.1.10",
-    "babel-eslint": "^3.0.1",
-    "babel-loader": "^5.2.1",
-    "bootstrap": "^3.3.4",
+    "babel": "^5.8.19",
+    "babel-core": "^5.8.19",
+    "babel-eslint": "^4.0.5",
+    "babel-loader": "^5.3.2",
+    "bootstrap": "^3.3.5",
     "brfs": "^1.4.0",
-    "chai": "^3.0.0",
-    "child-process-promise": "^1.0.1",
-    "codemirror": "^5.0.0",
-    "colors": "^1.0.3",
-    "css-loader": "^0.14.1",
-    "es5-shim": "^4.1.0",
-    "eslint": "0.24.0",
-    "eslint-plugin-babel": "^1.0.0",
+    "chai": "^3.2.0",
+    "child-process-promise": "^1.1.0",
+    "codemirror": "^5.5.0",
+    "colors": "^1.1.2",
+    "css-loader": "^0.15.6",
+    "es5-shim": "^4.1.10",
+    "eslint": "1.0.0",
+    "eslint-plugin-babel": "^1.2.0",
+    "eslint-plugin-lodash": "^0.1.3",
     "eslint-plugin-mocha": "^0.4.0",
-    "eslint-plugin-react": "^2.1.0",
-    "express": "^4.12.3",
-    "extract-text-webpack-plugin": "^0.8.0",
-    "file-loader": "^0.8.1",
-    "fs-extra": "^0.20.0",
+    "eslint-plugin-react": "^3.1.0",
+    "express": "^4.13.1",
+    "extract-text-webpack-plugin": "^0.8.2",
+    "file-loader": "^0.8.4",
+    "fs-extra": "^0.22.1",
     "fs-promise": "^0.3.1",
-    "glob": "^5.0.10",
+    "glob": "^5.0.14",
     "http-proxy": "^1.11.1",
-    "ip": "^0.3.2",
+    "ip": "^0.3.3",
     "isparta-loader": "^0.2.0",
-    "json-loader": "^0.5.1",
-    "karma": "~0.12.32",
-    "karma-chai": "^0.1.0",
+    "json-loader": "^0.5.2",
+    "karma": "~0.13.3",
     "karma-chrome-launcher": "~0.2.0",
     "karma-cli": "0.1.0",
     "karma-coverage": "^0.4.2",
-    "karma-coveralls": "^1.1.0",
-    "karma-firefox-launcher": "~0.1.3",
+    "karma-coveralls": "^1.1.2",
+    "karma-firefox-launcher": "~0.1.6",
     "karma-mocha": "~0.2.0",
-    "karma-mocha-reporter": "^1.0.2",
+    "karma-mocha-reporter": "^1.0.4",
     "karma-phantomjs-launcher": "^0.2.0",
-    "karma-sinon": "^1.0.3",
     "karma-sinon-chai": "^1.0.0",
-    "karma-sourcemap-loader": "^0.3.4",
-    "karma-webpack": "^1.5.0",
-    "less": "^2.4.0",
-    "less-loader": "^2.1.0",
-    "lodash": "^3.5.0",
-    "mocha": "^2.2.1",
-    "mt-changelog": "^0.5.2",
+    "karma-sourcemap-loader": "^0.3.5",
+    "karma-webpack": "^1.7.0",
+    "less": "^2.5.1",
+    "less-loader": "^2.2.0",
+    "marked": "^0.3.5",
+    "mocha": "^2.2.5",
+    "mt-changelog": "^0.6.1",
     "node-libs-browser": "^0.5.2",
-    "nodemon": "^1.3.7",
-    "output-file-sync": "^1.1.0",
+    "nodemon": "^1.4.0",
+    "output-file-sync": "^1.1.1",
     "phantomjs": "^1.9.17",
     "portfinder": "^0.4.0",
-    "react": "^0.13.1",
-    "react-component-metadata": "^1.1.1",
-    "react-hot-loader": "^1.2.7",
-    "react-router": "^0.13.1",
-    "rimraf": "^2.3.2",
-    "semver": "^4.3.4",
-    "sinon": "^1.10.3",
-    "sinon-chai": "^2.7.0",
-    "style-loader": "^0.12.0",
-    "transform-loader": "^0.2.1",
-    "webpack": "^1.7.2",
-    "webpack-dev-server": "^1.8.2",
-    "yargs": "^3.5.4"
+    "react": "^0.13.3",
+    "react-component-metadata": "^1.2.1",
+    "react-hot-loader": "^1.2.8",
+    "react-router": "^0.13.3",
+    "rimraf": "^2.4.2",
+    "semver": "^5.0.1",
+    "sinon": "^1.15.4",
+    "sinon-chai": "^2.8.0",
+    "style-loader": "^0.12.3",
+    "transform-loader": "^0.2.2",
+    "webpack": "^1.10.5",
+    "webpack-dev-server": "^1.10.1",
+    "yargs": "^3.16.1"
   },
   "dependencies": {
-    "classnames": "^2.0.0"
+    "babel-runtime": "^5.8.19",
+    "lodash": "^3.10.0",
+    "classnames": "^2.1.3"
   }
 }
