@@ -2,6 +2,7 @@ import React from 'react';
 import classNames from 'classnames';
 import BootstrapMixin from './BootstrapMixin';
 import CustomPropTypes from './utils/CustomPropTypes';
+import ButtonInput from './ButtonInput';
 
 const Button = React.createClass({
   mixins: [BootstrapMixin],
@@ -12,16 +13,23 @@ const Button = React.createClass({
     block: React.PropTypes.bool,
     navItem: React.PropTypes.bool,
     navDropdown: React.PropTypes.bool,
+    /**
+     * You can use a custom element for this component
+     */
     componentClass: CustomPropTypes.elementType,
     href: React.PropTypes.string,
-    target: React.PropTypes.string
+    target: React.PropTypes.string,
+    /**
+     * Defines HTML button type Attribute
+     * @type {("button"|"reset"|"submit")}
+     */
+    type: React.PropTypes.oneOf(ButtonInput.types)
   },
 
   getDefaultProps() {
     return {
       bsClass: 'button',
-      bsStyle: 'default',
-      type: 'button'
+      bsStyle: 'default'
     };
   },
 
@@ -68,6 +76,7 @@ const Button = React.createClass({
     return (
       <Component
         {...this.props}
+        type={this.props.type || 'button'}
         className={classNames(this.props.className, classes)}>
         {this.props.children}
       </Component>
