@@ -5,7 +5,7 @@ function renderDropdownButton (title, i) {
     <DropdownButton bsStyle={title.toLowerCase()} title={title} key={i}>
       <MenuItem eventKey='1'>Action</MenuItem>
       <MenuItem eventKey='2'>Another action</MenuItem>
-      <MenuItem eventKey='3' active={true}>Active Item</MenuItem>
+      <MenuItem eventKey='3' active>Active Item</MenuItem>
       <MenuItem divider />
       <MenuItem eventKey='4'>Separated link</MenuItem>
     </DropdownButton>
