@@ -9,13 +9,13 @@ const NavJustified = React.createClass({
         <Nav bsStyle='tabs' justified activeKey={1} onSelect={this.handleSelect}>
           <NavItem eventKey={1} href='/home'>NavItem 1 content</NavItem>
           <NavItem eventKey={2} title='Item'>NavItem 2 content</NavItem>
-          <NavItem eventKey={3} disabled={true}>NavItem 3 content</NavItem>
+          <NavItem eventKey={3} disabled>NavItem 3 content</NavItem>
         </Nav>
         <br />
         <Nav bsStyle='pills' justified activeKey={1} onSelect={this.handleSelect}>
           <NavItem eventKey={1} href='/home'>NavItem 1 content</NavItem>
           <NavItem eventKey={2} title='Item'>NavItem 2 content</NavItem>
-          <NavItem eventKey={3} disabled={true}>NavItem 3 content</NavItem>
+          <NavItem eventKey={3} disabled>NavItem 3 content</NavItem>
         </Nav>
       </div>
     );
