@@ -1,5 +1,9 @@
+function handleClick() {
+  alert('You have clicked on me');
+}
+
 const panelInstance = (
-  <Panel>
+  <Panel onClick={ handleClick }>
     Basic panel example
   </Panel>
 );
