@@ -1,13 +1,33 @@
-const positionerInstance = (
-  <ButtonToolbar>
-    <OverlayTrigger
-      container={mountNode} containerPadding={20}
-      trigger='click' placement='bottom'
-      overlay={<Popover title='Popover bottom'><strong>Holy guacamole!</strong> Check this info.</Popover>}
-    >
-      <Button bsStyle='default'>Holy guacamole!</Button>
-    </OverlayTrigger>
-  </ButtonToolbar>
-);
+class Example extends React.Component {
+  constructor(...args){
+    super(...args);
+    this.state = { show: false };
+  }
+  render(){
 
-React.render(positionerInstance, mountNode);
+    return (
+      <ButtonToolbar>
+        <Button
+          bsStyle='default'
+          onClick={e => this.setState({ target: e.target, show: !this.state.show })}
+        >
+          Holy guacamole!
+        </Button>
+
+        <Overlay
+          show={this.state.show}
+          target={()=> React.findDOMNode(this.state.target)}
+          placement='bottom'
+          container={mountNode}
+          containerPadding={20}
+        >
+          <Popover title='Popover bottom'>
+            <strong>Holy guacamole!</strong> Check this info.
+          </Popover>
+        </Overlay>
+      </ButtonToolbar>
+    );
+  }
+}
+
+React.render(<Example/>, mountNode);
