@@ -1,8 +1,8 @@
 const progressInstance = (
   <ProgressBar>
-    <ProgressBar bsStyle='success' now={35} key={1} />
+    <ProgressBar striped bsStyle='success' now={35} key={1} />
     <ProgressBar bsStyle='warning' now={20} key={2} />
-    <ProgressBar bsStyle='danger' now={10} key={3} />
+    <ProgressBar active bsStyle='danger' now={10} key={3} />
   </ProgressBar>
 );
 
