@@ -4,7 +4,7 @@ import PanelGroup from './PanelGroup';
 const Accordion = React.createClass({
   render() {
     return (
-      <PanelGroup {...this.props} accordion={true}>
+      <PanelGroup {...this.props} accordion>
         {this.props.children}
       </PanelGroup>
     );
