@@ -1,15 +1,11 @@
 import React from 'react';
 import InputBase from './InputBase';
-import ButtonInput from './ButtonInput';
-import FormControls from './FormControls';
+import * as FormControls from './FormControls';
 import deprecationWarning from './utils/deprecationWarning';
 
 class Input extends InputBase {
   render() {
-    if (ButtonInput.types.indexOf(this.props.type) > -1) {
-      deprecationWarning(`Input type=${this.props.type}`, 'ButtonInput');
-      return <ButtonInput {...this.props} />;
-    } else if (this.props.type === 'static') {
+    if (this.props.type === 'static') {
       deprecationWarning('Input type=static', 'StaticText');
       return <FormControls.Static {...this.props} />;
     }
@@ -17,5 +13,9 @@ class Input extends InputBase {
     return super.render();
   }
 }
+
+Input.propTypes = {
+  type: React.PropTypes.string
+};
 
 export default Input;
