@@ -1,47 +1,70 @@
 import React from 'react';
 import ReactTestUtils from 'react/lib/ReactTestUtils';
 import Modal from '../src/Modal';
+import { render, shouldWarn } from './helpers';
 
 describe('Modal', function () {
+  let mountPoint;
+
+  beforeEach(()=>{
+    mountPoint = document.createElement('div');
+    document.body.appendChild(mountPoint);
+  });
+
+  afterEach(function () {
+    React.unmountComponentAtNode(mountPoint);
+    document.body.removeChild(mountPoint);
+  });
 
   it('Should render the modal content', function() {
     let noOp = function () {};
-    let instance = ReactTestUtils.renderIntoDocument(
-      <Modal onRequestHide={noOp}>
+    let instance = render(
+      <Modal show onHide={noOp} animation={false}>
         <strong>Message</strong>
       </Modal>
-    );
-    assert.ok(ReactTestUtils.findRenderedDOMComponentWithTag(instance, 'strong'));
+    , mountPoint);
+
+    assert.ok(
+      ReactTestUtils.findRenderedDOMComponentWithTag(instance.refs.modal, 'strong'));
   });
 
   it('Should add modal-open class to the modal container while open', function(done) {
 
     let Container = React.createClass({
       getInitialState() {
-        return {modalOpen: true};
+        return { modalOpen: true };
       },
       handleCloseModal() {
-        this.setState({modalOpen: false});
+        this.setState({ modalOpen: false });
       },
       render() {
-        if (this.state.modalOpen) {
-          return (
-            <Modal onRequestHide={this.handleCloseModal} container={this}>
+        return (
+          <div>
+            <Modal
+              animation={false}
+              show={this.state.modalOpen}
+              onHide={this.handleCloseModal}
+              container={this}
+            >
               <strong>Message</strong>
             </Modal>
-          );
-        } else {
-          return <span/>;
-        }
+          </div>
+        );
       }
     });
-    let instance = ReactTestUtils.renderIntoDocument(
-        <Container />
-    );
-    assert.ok(React.findDOMNode(instance).className.match(/\modal-open\b/));
 
-    let backdrop = React.findDOMNode(instance).getElementsByClassName('modal-backdrop')[0];
+    let instance = render(
+          <Container />
+        , mountPoint);
+
+    let modal = ReactTestUtils.findRenderedComponentWithType(instance, Modal);
+
+    assert.ok(React.findDOMNode(instance).className.match(/\bmodal-open\b/));
+
+    let backdrop = React.findDOMNode(modal.refs.backdrop);
+
     ReactTestUtils.Simulate.click(backdrop);
+
     setTimeout(function(){
       assert.equal(React.findDOMNode(instance).className.length, 0);
       done();
@@ -51,55 +74,180 @@ describe('Modal', function () {
 
   it('Should close the modal when the backdrop is clicked', function (done) {
     let doneOp = function () { done(); };
-    let instance = ReactTestUtils.renderIntoDocument(
-      <Modal onRequestHide={doneOp}>
+    let instance = render(
+      <Modal show onHide={doneOp} animation={false}>
         <strong>Message</strong>
       </Modal>
-    );
+    , mountPoint);
 
-    let backdrop = React.findDOMNode(instance).getElementsByClassName('modal-backdrop')[0];
+    let backdrop = React.findDOMNode(instance.refs.backdrop);
+
     ReactTestUtils.Simulate.click(backdrop);
   });
 
-  it('Should close the modal when the modal background is clicked', function (done) {
+  it('Should close the modal when the modal dialog is clicked', function (done) {
     let doneOp = function () { done(); };
-    let instance = ReactTestUtils.renderIntoDocument(
-      <Modal onRequestHide={doneOp}>
+
+    let instance = render(
+      <Modal show onHide={doneOp}>
         <strong>Message</strong>
       </Modal>
-    );
+    , mountPoint);
 
-    let backdrop = React.findDOMNode(instance).getElementsByClassName('modal')[0];
+    let dialog = React.findDOMNode(instance.refs.dialog);
+
+    ReactTestUtils.Simulate.click(dialog);
+  });
+
+  it('Should not close the modal when the "static" backdrop is clicked', function () {
+    let onHideSpy = sinon.spy();
+    let instance = render(
+      <Modal show onHide={onHideSpy} backdrop='static'>
+        <strong>Message</strong>
+      </Modal>
+    , mountPoint);
+
+    let backdrop = React.findDOMNode(instance.refs.backdrop);
+
     ReactTestUtils.Simulate.click(backdrop);
+
+    expect(onHideSpy).to.not.have.been.called;
+  });
+
+  it('Should close the modal when the modal close button is clicked', function (done) {
+    let doneOp = function () { done(); };
+
+    let instance = render(
+      <Modal show onHide={doneOp}>
+        <Modal.Header closeButton />
+        <strong>Message</strong>
+      </Modal>
+    , mountPoint);
+
+    let button = React.findDOMNode(instance.refs.modal)
+        .getElementsByClassName('close')[0];
+
+    ReactTestUtils.Simulate.click(button);
+  });
+
+  it('Should use bsClass on the dialog', function () {
+    let noOp = function () {};
+    let instance = render(
+      <Modal show bsClass='mymodal' onHide={noOp}>
+        <strong>Message</strong>
+      </Modal>
+    , mountPoint);
+
+    let dialog = React.findDOMNode(instance.refs.dialog);
+    let backdrop = React.findDOMNode(instance.refs.backdrop);
+
+    assert.ok(dialog.className.match(/\bmymodal\b/));
+    assert.ok(dialog.children[0].className.match(/\bmymodal-dialog\b/));
+    assert.ok(dialog.children[0].children[0].className.match(/\bmymodal-content\b/));
+
+    assert.ok(backdrop.className.match(/\bmymodal-backdrop\b/));
+
+
+    shouldWarn("Invalid prop 'bsClass' of value 'mymodal'");
   });
 
   it('Should pass bsSize to the dialog', function () {
     let noOp = function () {};
-    let instance = ReactTestUtils.renderIntoDocument(
-      <Modal bsSize='small' onRequestHide={noOp}>
+    let instance = render(
+      <Modal show bsSize='small' onHide={noOp}>
         <strong>Message</strong>
       </Modal>
-    );
+    , mountPoint);
 
-    let dialog = React.findDOMNode(instance).getElementsByClassName('modal-dialog')[0];
+    let dialog = React.findDOMNode(instance.refs.modal).getElementsByClassName('modal-dialog')[0];
     assert.ok(dialog.className.match(/\bmodal-sm\b/));
   });
 
   it('Should pass dialogClassName to the dialog', function () {
     let noOp = function () {};
-    let instance = ReactTestUtils.renderIntoDocument(
-      <Modal dialogClassName="testCss" onRequestHide={noOp}>
+    let instance = render(
+      <Modal show dialogClassName="testCss" onHide={noOp}>
         <strong>Message</strong>
       </Modal>
-    );
+    , mountPoint);
 
-    let dialog = ReactTestUtils.findRenderedDOMComponentWithClass(instance, 'modal-dialog');
+    let dialog = ReactTestUtils.findRenderedDOMComponentWithClass(instance.refs.modal, 'modal-dialog');
     assert.match(dialog.props.className, /\btestCss\b/);
+  });
+
+  it('Should assign refs correctly when no backdrop', function () {
+
+    let test = () => render(
+      <Modal show backdrop={false} onHide={function () {}}>
+        <strong>Message</strong>
+      </Modal>
+    , mountPoint);
+
+    expect(test).not.to.throw();
+  });
+
+  it('Should use dialogComponent', function () {
+    let noOp = function () {};
+
+    class CustomDialog {
+      render(){ return <div {...this.props}/>; }
+    }
+
+    let instance = render(
+      <Modal show dialogComponent={CustomDialog} onHide={noOp}>
+        <strong>Message</strong>
+      </Modal>
+    , mountPoint);
+
+    assert.ok(instance.refs.dialog instanceof CustomDialog);
+  });
+
+  it('Should pass transition callbacks to Transition', function (done) {
+    let count = 0;
+    let increment = ()=> count++;
+
+    let instance = render(
+      <Modal show
+        onHide={()=>{}}
+        onExit={increment}
+        onExiting={increment}
+        onExited={()=> {
+          increment();
+          expect(count).to.equal(6);
+          done();
+        }}
+        onEnter={increment}
+        onEntering={increment}
+        onEntered={()=> {
+          increment();
+          instance.setProps({ show: false });
+        }}
+      >
+        <strong>Message</strong>
+      </Modal>
+      , mountPoint);
+  });
+
+  it('Should unbind listeners when unmounted', function() {
+    render(
+        <div>
+          <Modal show onHide={() => null} animation={false}>
+            <strong>Foo bar</strong>
+          </Modal>
+        </div>
+    , mountPoint);
+
+    assert.include(document.body.className, 'modal-open');
+
+    render(<div />, mountPoint);
+
+    assert.notInclude(document.body.className, 'modal-open');
   });
 
   describe('Focused state', function () {
     let focusableContainer = null;
-    beforeEach(function () {
+
+    beforeEach(()=>{
       focusableContainer = document.createElement('div');
       focusableContainer.tabIndex = 0;
       document.body.appendChild(focusableContainer);
@@ -111,113 +259,48 @@ describe('Modal', function () {
       document.body.removeChild(focusableContainer);
     });
 
-    it('Should focus on the Modal when it is opened', function (done) {
+    it('Should focus on the Modal when it is opened', function () {
+
       document.activeElement.should.equal(focusableContainer);
 
-      let doneOp = function () {
-        // focus should be back on the previous element when modal closed
-        setTimeout(function () {
-          document.activeElement.should.equal(focusableContainer);
-          done();
-        }, 0);
-      };
+      let instance = render(
+        <Modal show onHide={()=>{}} animation={false}>
+          <strong>Message</strong>
+        </Modal>
+        , focusableContainer);
 
-      let Container = React.createClass({
-        getInitialState() {
-          return {modalOpen: true};
-        },
-        handleCloseModal() {
-          this.setState({modalOpen: false});
-          doneOp();
-        },
-        render() {
-          if (this.state.modalOpen) {
-            return (
-              <Modal onRequestHide={this.handleCloseModal} container={this}>
-                <strong>Message</strong>
-              </Modal>
-            );
-          } else {
-            return <span/>;
-          }
-        }
-      });
+      document.activeElement.className.should.contain('modal');
 
-      let instance = React.render(<Container />, focusableContainer);
+      instance.renderWithProps({ show: false });
 
-      setTimeout(function () {
-        // modal should be focused when opened
-        let modal = React.findDOMNode(instance).getElementsByClassName('modal')[0];
-        document.activeElement.should.equal(modal);
-
-        // close the modal
-        let backdrop = React.findDOMNode(instance).getElementsByClassName('modal-backdrop')[0];
-        ReactTestUtils.Simulate.click(backdrop);
-      }, 0);
+      document.activeElement.should.equal(focusableContainer);
     });
 
-    it('Should not focus on the Modal when autoFocus is false', function (done) {
+
+    it('Should not focus on the Modal when autoFocus is false', function () {
+      render(
+        <Modal show autoFocus={false} onHide={()=>{}} animation={false}>
+          <strong>Message</strong>
+        </Modal>
+        , focusableContainer);
 
       document.activeElement.should.equal(focusableContainer);
-
-      let Container = React.createClass({
-        getInitialState() {
-          return {modalOpen: true};
-        },
-        render() {
-          if (this.state.modalOpen) {
-            return (
-              <Modal autoFocus={false} onRequestHide={()=>{}} container={this}>
-                <strong>Message</strong>
-              </Modal>
-            );
-          } else {
-            return <span/>;
-          }
-        }
-      });
-
-      React.render(<Container />, focusableContainer);
-
-      setTimeout(function () {
-        // modal should be focused when opened
-        document.activeElement.should.equal(focusableContainer);
-        done();
-      }, 0);
     });
 
-    it('Should not focus Modal when child has focus', function (done) {
+    it('Should not focus Modal when child has focus', function () {
 
       document.activeElement.should.equal(focusableContainer);
 
-      let Container = React.createClass({
-        getInitialState() {
-          return {modalOpen: true};
-        },
-        render() {
-          if (this.state.modalOpen) {
-            return (
-              <Modal onRequestHide={()=>{}} container={this}>
-                <input autoFocus />
-              </Modal>
-            );
-          } else {
-            return <span/>;
-          }
-        }
-      });
+      render(
+        <Modal show onHide={()=>{}} animation={false}>
+          <input autoFocus />
+        </Modal>
+        , focusableContainer);
 
-      var instance = React.render(<Container />, focusableContainer);
+      let input = document.getElementsByTagName('input')[0];
 
-      setTimeout(function () {
-        let input = React.findDOMNode(
-          ReactTestUtils.findRenderedDOMComponentWithTag(instance, 'input'))
-
-        document.activeElement.should.equal(input);
-        done();
-      }, 0);
+      document.activeElement.should.equal(input);
     });
   });
-
 
 });
