@@ -83,9 +83,9 @@ describe('Nav', function () {
           </Nav>
         );
 
-    let items = ReactTestUtils.scryRenderedComponentsWithType(instance, NavItem);
+    let items = ReactTestUtils.scryRenderedDOMComponentsWithTag(instance, 'A');
 
-    ReactTestUtils.Simulate.click(items[1].refs.anchor);
+    ReactTestUtils.Simulate.click(items[1]);
   });
 
   it('Should set the correct item active by href', function () {
@@ -113,6 +113,68 @@ describe('Nav', function () {
     let items = ReactTestUtils.scryRenderedComponentsWithType(instance, NavItem);
 
     assert.ok(items[0].props.navItem);
+  });
+
+  it('Should apply className only to the wrapper nav element', function () {
+    const instance = ReactTestUtils.renderIntoDocument(
+      <Nav bsStyle="tabs" activeKey={1} className="nav-specific">
+        <NavItem key={1}>Tab 1 content</NavItem>
+        <NavItem key={2}>Tab 2 content</NavItem>
+      </Nav>
+    );
+
+    let ulNode = React.findDOMNode(ReactTestUtils.findRenderedDOMComponentWithTag(instance, 'ul'));
+    assert.notInclude(ulNode.className, 'nav-specific');
+
+    let navNode = React.findDOMNode(ReactTestUtils.findRenderedDOMComponentWithTag(instance, 'nav'));
+    assert.include(navNode.className, 'nav-specific');
+  });
+
+  it('Should apply ulClassName to the inner ul element', function () {
+    const instance = ReactTestUtils.renderIntoDocument(
+      <Nav bsStyle="tabs" activeKey={1} className="nav-specific" ulClassName="ul-specific">
+        <NavItem key={1}>Tab 1 content</NavItem>
+        <NavItem key={2}>Tab 2 content</NavItem>
+      </Nav>
+    );
+
+    let ulNode = React.findDOMNode(ReactTestUtils.findRenderedDOMComponentWithTag(instance, 'ul'));
+    assert.include(ulNode.className, 'ul-specific');
+    assert.notInclude(ulNode.className, 'nav-specific');
+
+    let navNode = React.findDOMNode(ReactTestUtils.findRenderedDOMComponentWithTag(instance, 'nav'));
+    assert.notInclude(navNode.className, 'ul-specific');
+    assert.include(navNode.className, 'nav-specific');
+  });
+
+  it('Should apply id to the wrapper nav element', function () {
+    const instance = ReactTestUtils.renderIntoDocument(
+      <Nav bsStyle="tabs" activeKey={1} id="nav-id">
+        <NavItem key={1}>Tab 1 content</NavItem>
+        <NavItem key={2}>Tab 2 content</NavItem>
+      </Nav>
+    );
+
+    let navNode = React.findDOMNode(ReactTestUtils.findRenderedDOMComponentWithTag(instance, 'nav'));
+    assert.equal(navNode.id, 'nav-id');
+
+    let ulNode = React.findDOMNode(ReactTestUtils.findRenderedDOMComponentWithTag(instance, 'ul'));
+    assert.notEqual(ulNode.id, 'nav-id');
+  });
+
+  it('Should apply ulId to the inner ul element', function () {
+    const instance = ReactTestUtils.renderIntoDocument(
+      <Nav bsStyle="tabs" activeKey={1} id="nav-id" ulId="ul-id">
+        <NavItem key={1}>Tab 1 content</NavItem>
+        <NavItem key={2}>Tab 2 content</NavItem>
+      </Nav>
+    );
+
+    let ulNode = React.findDOMNode(ReactTestUtils.findRenderedDOMComponentWithTag(instance, 'ul'));
+    assert.equal(ulNode.id, 'ul-id');
+
+    let navNode = React.findDOMNode(ReactTestUtils.findRenderedDOMComponentWithTag(instance, 'nav'));
+    assert.equal(navNode.id, 'nav-id');
   });
 
 
