@@ -5,20 +5,25 @@ import Tooltip from '../src/Tooltip';
 describe('Tooltip', function () {
   it('Should output a tooltip with content', function () {
     let instance = ReactTestUtils.renderIntoDocument(
-      <Tooltip>
+      <Tooltip positionTop={10} positionLeft={20}>
         <strong>Tooltip Content</strong>
       </Tooltip>
     );
     assert.ok(ReactTestUtils.findRenderedDOMComponentWithTag(instance, 'strong'));
-    assert.ok(ReactTestUtils.findRenderedDOMComponentWithClass(instance, 'fade'));
+
+    const tooltip = ReactTestUtils.findRenderedDOMComponentWithClass(instance, 'tooltip');
+    assert.deepEqual(tooltip.props.style, {top: 10, left: 20});
   });
 
-  it('Should not have the fade class if animation is false', function () {
-    let instance = ReactTestUtils.renderIntoDocument(
-      <Tooltip animation={false}>
-        <strong>Tooltip Content</strong>
-      </Tooltip>
-    );
-    assert.equal(React.findDOMNode(instance).className.match(/\bfade\b/), null, 'The fade class should not be present');
+  describe('When a style property is provided', function () {
+    it('Should render a tooltip with merged styles', function () {
+      let instance = ReactTestUtils.renderIntoDocument(
+        <Tooltip style={{opacity: 0.9}} positionTop={10} positionLeft={20}>
+          <strong>Tooltip Content</strong>
+        </Tooltip>
+      );
+      const tooltip = ReactTestUtils.findRenderedDOMComponentWithClass(instance, 'tooltip');
+      assert.deepEqual(tooltip.props.style, {opacity: 0.9, top: 10, left: 20});
+    });
   });
 });
