@@ -3,23 +3,13 @@ import path from 'path';
 import fsp from 'fs-promise';
 import { copy } from '../fs-utils';
 import { exec } from '../exec';
-import generateFactories from '../generateFactories';
-import { repoRoot, srcRoot, bowerRoot } from '../constants';
-import { buildFolder } from '../buildBabel';
+import { repoRoot, bowerRoot } from '../constants';
 
 const packagePath = path.join(repoRoot, 'package.json');
 const bowerTemplate = path.join(__dirname, 'bower.json');
 const bowerJson = path.join(bowerRoot, 'bower.json');
 
 const readme = path.join(__dirname, 'README.md');
-const license = path.join(repoRoot, 'LICENSE');
-
-const babelOptions = {
-  __reactBootstrapDeprecationWarning: true,
-  modules: 'amd'
-};
-
-const libDestination = path.join(bowerRoot, 'lib');
 
 function bowerConfig() {
   return Promise.all([
@@ -36,13 +26,10 @@ export default function BuildBower() {
   console.log('Building: '.cyan + 'bower module'.green);
 
   return exec(`rimraf ${bowerRoot}`)
-    .then(() => fsp.mkdirs(libDestination))
+    .then(() => fsp.mkdirs(bowerRoot))
     .then(() => Promise.all([
       bowerConfig(),
-      generateFactories(libDestination, babelOptions),
-      buildFolder(srcRoot, libDestination, babelOptions),
-      copy(readme, bowerRoot),
-      copy(license, bowerRoot)
+      copy(readme, bowerRoot)
     ]))
     .then(() => console.log('Built: '.cyan + 'bower module'.green));
 }
