@@ -6,21 +6,21 @@ import { copy } from './fs-utils';
 import { distRoot, bowerRoot } from './constants';
 import { exec } from './exec';
 
-function forkAndBuildDocs(verbose) {
+function forkAndBuildDocs({verbose}) {
   console.log('Building: '.cyan + 'docs'.green);
 
-  let options = verbose ? ' -- --verbose' : '';
+  const verboseOption = verbose ? '--verbose' : '';
 
-  return exec(`npm run docs-build${options}`)
+  return exec(`npm run docs-build -- ${verboseOption}`)
     .then(() => console.log('Built: '.cyan + 'docs'.green));
 }
 
-export default function Build(verbose) {
+export default function Build(options) {
   return Promise.all([
       lib(),
       bower(),
       dist(),
-      forkAndBuildDocs(verbose)
+      forkAndBuildDocs(options)
     ])
     .then(() => copy(distRoot, bowerRoot));
 }
