@@ -16,6 +16,7 @@ import { bowerRepo, bowerRoot, tmpBowerRepo, docsRoot, docsRepo, tmpDocsRepo } f
 
 const yargsConf = yargs
   .usage('Usage: $0 <version> [--preid <identifier>]\nor\nUsage: $0 --docs')
+  .help('h')
   .example('$0 --docs', 'Release only docs')
   .example('$0 minor --preid beta', 'Release with minor version bump with pre-release tag')
   .example('$0 major', 'Release with major version bump')
@@ -38,7 +39,7 @@ const yargsConf = yargs
     alias: 'n',
     demand: false,
     default: false,
-    describe: 'Execute command in dry run mode. Will not commit, tag, push, or publish anything. Userful for testing.'
+    describe: 'Execute command in dry run mode. Will not commit, tag, push, or publish anything. Useful for testing.'
   })
   .option('verbose', {
     demand: false,
